+++ src/vmm/src/devices/virtio/block/test_utils.rs
@@ -166,9 +166,9 @@ impl<'a, 'b> RequestDescriptorChain<'a, 'b> {
 /// The head of the chain is made available as the first descriptor to be processed, by
 /// setting avail_idx to 1.
 pub fn read_blk_req_descriptors(vq: &VirtQueue) {
-    let request_type_desc: usize = 0;
-    let data_desc: usize = 1;
-    let status_desc: usize = 2;
+    let request_type_desc: u16 = 0;
+    let data_desc: u16 = 1;
+    let status_desc: u16 = 2;
 
     let request_addr: u64 = 0x1000;
     let data_addr: u64 = 0x2000;
@@ -176,26 +176,21 @@ pub fn read_blk_req_descriptors(vq: &VirtQueue) {
     let len = 0x1000;
 
     // Set the request type descriptor.
-    vq.avail.ring[request_type_desc].set(request_type_desc as u16);
-    vq.dtable[request_type_desc].set(request_addr, len, VIRTQ_DESC_F_NEXT, data_desc as u16);
+    vq.avail.ring[request_type_desc as usize].set(request_type_desc);
+    vq.dtable[request_type_desc as usize].set(request_addr, len, VIRTQ_DESC_F_NEXT, data_desc);
 
     // Set the data descriptor.
-    vq.avail.ring[data_desc].set(data_desc as u16);
-    vq.dtable[data_desc].set(
+    vq.avail.ring[data_desc as usize].set(data_desc);
+    vq.dtable[data_desc as usize].set(
         data_addr,
         len,
         VIRTQ_DESC_F_NEXT | VIRTQ_DESC_F_WRITE,
-        status_desc as u16,
+        status_desc,
     );
 
     // Set the status descriptor.
-    vq.avail.ring[status_desc].set(status_desc as u16);
-    vq.dtable[status_desc].set(
-        status_addr,
-        len,
-        VIRTQ_DESC_F_WRITE,
-        (status_desc + 1) as u16,
-    );
+    vq.avail.ring[status_desc as usize].set(status_desc);
+    vq.dtable[status_desc as usize].set(status_addr, len, VIRTQ_DESC_F_WRITE, status_desc + 1);
 
     // Mark the next available descriptor.
     vq.avail.idx.set(1);
